Fix mapped unsubscribe affecting other subscribers
This fixes an issue where if a mapped atom is subscribed multiple times, then unsubscribed by one, it would effectively unsubscribe the others because the `listener` callback would only get added to the parent atom's listener set once. Now it'll only get removed if there are no subscribers to the derived atom remaining.